Read and Cache Tables

Creates a cache of database tables for efficient access. Tables are read from the database and stored in a list.

Usage

read_and_cache(db_ops, dynamic_config)

Arguments

db_ops

Database operations object that contains methods for reading tables. Must have a read_table method that accepts a table name as parameter.

dynamic_config

List of table configurations. Each configuration must be a list containing at least a table_name field specifying which table to read.

Value

A named list of cached tables. Access individual tables using tables$table_name.

Examples

if (FALSE) { # \dontrun{
# In server.R
config <- list(
  list(
    table_name = "config_packages",
    group_type = "choice",
    group_col = "package",
    choices_col = "version"
  ),
  list(
    table_name = "config_pid",
    group_type = "param",
    group_col = "pid",
    display_col = "full_name"
  )
)
tables <- read_and_cache(db_ops, config)

# Access cached table
output$packageTable <- renderTable({
  tables$config_packages
})
} # }